Catalycity update and sign fix
With this refactoring, the catalycity for the wall is only a
function of the reactant gas. Each of the CatalyticWallBase
children will handle supplying both reactant and product wall
fluxes through the apply_fluxes function. Thus, no need to build
a product catalycity object anymore. Also, previously the sign was
built into the CatalycityBase object. This was a bad idea. Removed
and now the CatalyticWallBase and children will worry about who's
positive and negative.