Add accessor for MeshBuilder from SimulationBuilder
This is needed for doing post-input refinement when there's
a restart file. We should really refactor SimulationBuilder to
just have accessors for each of the builder modules. I still like
containing everything in SimulationBuilder to keep the Simulation
API simple, but redoing function calls is stupid and everything
in the builder modules should be const anyway.