Documenting alternative GNU Make style for grabbing files
This is in case the AC_SUBST version breaks and we need to do
something else. In theory, the AC_SUBST version is more portable
because it doesn't rely on GNU Make, but it is relying on some
autoconf foo that wouldn't surprise me if it broke. So, we'll
have this here to fall back on