Remove docs from EXTRA_DIST target
Having docs there would cause Doxygen to be run when `make dist`
is run. However, the docs target was removed from the Doxygen target
list, so this was giving me trouble with make dist(check). We probably
don't want to distribute Doxygen stuff anyway and should get around
to setting up the webpage with Doxygen documentation...