Refactor Multiphysics system, add nonlocal_foo()
This is going to be necessary to get AveragedTurbine working, but it's
a nice code reduction too.

We could make these method pointers into template parameters instead
to make the generated code equivalent to the previous case, but this
code is already heavyweight enough that funky templating is not where
I'd start with optimization.