Update init interface for SolidMechancs variables.
We need this for cases where the elements are manifolds in space.
libMesh::Elem::mesh_dimension() only reports the manifold dimension
and libMesh::MeshBase::spatial_dimension is a compile time constant.
Thus, we need to provide an interface to initialize the displacement
variables that we'll need. Still have to think about to actually
determine the dimension at runtime...