Helpers in SolidMechanicsVariables to track if v and/or w are active
We need this because we can have domains that are lower dimensional
manifolds, but still have unknowns in the larger space. E.g. we can
a sheet in 2-D and 3-D even though it's a 2-D manifold. Similarly,
a 1-D bar can displace in 2 and 3 dimensions. But we can only determine
the manifold dimension, implicitly, at run time. Thus, we provide these
knobs to turn so that if the sheet is only in 2-D, we don't add extra
dofs for the z-displacement. This aspect not fully enabled yet in
ElasticMembrane, but this will provide the hooks for it.