Shut up warnings in BoussinesqBuoyancySPGSMStabilization
This is related to #93. In particular, I was playing with including
some additional terms in the Boussinesq buoyancy that pop out for
SPGSM/VMS stabilization, but they seemed to do weird things, so
left them commented out. The things that were getting unused warnings
are related to that so I left them in there commented out for whenever
someone has time to look at this.