Adding compute_postprocessing_quantity
This is the API we'll move to to have each Physics compute their
postprocessing quantities. I don't really see there being much gain
with dealing with all the Cache shit vs. just evaluating it at each
point. If the repeated calls to each Physics pops up in timing, then
we can revisit. But this is much simpler code.