Always have Doxygen build Antioch and Cantera
Before, we didn't have Doxygen picking up optional classes because
we didn't properly setup having it find grins_config.h in the build
directory. But, it turns out that we can just PREDEFINE having the
optional classes anyway so that they're always included in the Doxygen
documentation, whether or not those optional packages are actually
present in the build. I actually prefer this so that the documentation
is more complete without having to track down all the optional
dependencies.