No, this should be physics_name.
Contrary to 08e99bc49c27c8fdd6ab86336aa310fa05d39774, this really
should be physics_name because Stokes is subclassing this and
we should be able to parse Stokes section.

I actually think that the variable classes could actually be treated
as "mixins" and we use multiple inheritance in the root physics classes
with these variable classes.