Input file for new membrane+stiffener test
This is just a uniform refinement of the
elastic_mooney_rivlin_square_hookean_stiffeners_regression test.
We reuse the output file from that test as the initial guess since
the projected initial guess doesn't converge in the refined guess
(probably because we're not doing the projection correctly yet
on mixed dimensional grids).