[grins]: adding a new input parsing option
This is set in the input file to determine if the non-linear solver will continue after failures. It is set to 'false' by default.
This is accomplished using the current GRINS paradigm, so a bool type is specified as false by the constructor, unless overwritten by
explicitly parsing the option, 'continue_after_max_iterations' in the input file.