Penalty poiseuille regression test fixes
A multilevel fix here:

We move the pressure pinning point to the boundary where it's actually
applied, so we don't end up with a solver-dependent pressure offset.

We use ASM instead of BJ in parallel with the unstabilized case; some
overlap seems to be necessary to counteract the LBB-dubious subspaces
due to how libMesh assigns boundary dofs.

We add a new (eyeball-norm checked...) gold standard.