Make many properties ParameterUsers
Anything that fit the "Parameter has a name in the input file and gets
used directly without caching" model, I added.

These parameters do *not* get properly registered yet, though; for
that to work we first need to get the Physics classes to pass on
registration requests to their properties.