Cache elem id for PressurePinning
Here we change the logic of error detection for the pin location
in PressurePinning. Previously, we just cached a boolean and did a
contains_point() for every element. Now, we instead cache the element
id and initialize it to invalid_id in order to detect out-of-mesh
pin_locations. Now we can just compare element ids instead of calling
contains_point() in addition to getting error checking. Win win.

Also had to tweak the pin location in INS-Poiseuille test.
This is needed because the pin location was on the border between
two elements and changing the algorithm changed for picking out
the element and, in particular, we were hitting the noise tolerance
in the solution when a different element was grabbed.