Add MaterialsParsing::turb_viscosity_model
The way we'll do this is, upstream of the this function, a turbulent
viscosity model (e.g. SpallartAllmaras) is detected and we need to
look for the "physical" viscosity model. This function will then look
in [Materials/MATERIAL_NAME/Viscosity/turb_visc_model] for this
physical viscosity model and return it.