Only one 'rho' for Boussinesq Physics classes
There were multiple densities floating around between the Boussinesq
Physics classes which this commit removes. By assumption, density is
constant so we shouldn't be needing differing densities between
the fluid terms in the stabilization classes and the terms in the
"core" Boussinesq classes. So everyone now uses the density in the
base class. Also renamed the variable for consistency (not sure where
the "reference" came from, it's just density). This will setup the
material-based parsing changes.