Add material name to thermochemistry wrapper constructors
We're doing it this way instead of having the thermochemistry wrappers
find the name themselves because we need the physics name to get the
material name from the input file. So, instead of tying the wrappers
to a particular physics (or passing in the physics name), we just pass
in the material name.