Use variable classes in AxisymmetricHeatTransfer
and AxisymmetricBoussinesqBuoyancy. Note the order of the physics
and variable initialization was important here because the "gold"
restart had lumped the {u,v,T} variables together. First cut had
them separated and this was tripping an assert on the compute error.
So, somewhere in libMesh, there's a problem with trying to compare
systems with the variables grouped differently.