Deprecate unsteady-solver/transient option
Instead, what we're going to is look for SolverOptions/TimeStepping/solver_type.
This is going to require the user to set what solver_type they actually
want for timestepping (I think this is a good thing, we'll see what others
say). Right now, we only have libMesh::EulerSolver, but this will make
it easier to add others later. This new option is actually being parsed
yet, that's coming.