Add new "generic" test that leverages an exact solution
This test will mandate the user supply an exact solution for each of
the variables requested for error computation. The error computed
will be the error between the computed solution and the exact
solution in the specified norms. Currently, only L2 is supported.
However, what the test checks is that the difference between the
expected error, provided at the command line, and the computed error
is within the specified tolerance.

This test is similar in spirit to the generic_solution_code that
uses a "gold standard" data file to compute the error.