Static methods replacing globals for physics names
This is fixing one of the oldest gaffe's in GRINS. Instead of using
global namespaced variables for physics names, now use a class,
called PhysicsNaming, that has static methods. In addition to getting
Doxygen to pick this up now, this will setup the naming scheme to use
multiple instances of the same Physics in the input file.