Utility function for physics consistency checking
We can't assume that each suffix should match for each physics.
So, for doing the consistency checking on the physics, we'll
just strip the suffix off and looking for the "raw" required physics
in the list. This is a linear search, but shouldn't really effect
timings at all unless the PhysicsList gets pretty big.