Enable physics suffix parsing
The idea here is that each time we are going to build a physics,
we extract the suffix the user supplied and set it in PhysicsNaming.
Now every call to a physics name method will inherit the suffix so
that the suffix will propagate as needed for all the Physics to
parse the options corresponding to the physics+suffix the user set.