Add/use T/P/Y cp/cv interface to Antioch wrappers
Need a uniform interface for all third party thermochemistry+transport
libraries. (Really, this should've been a CRTP thing, but I was a noob.)
For Cantera, we always have to set the thermodynamic state each time,
so we need to provide the necessary information. This is ugly, but
I don't have anything better at present.