Add ozone flame example
This is a "confined" ozone flame, ignited by a hot wall. The example
is akin to one in Becker, Braack, and Rannacher, 1999; no details are
really given on the example, other than the mechanism, so it's just a
stab. Note that currently, Cantera's kinetics is apparently not
igniting, but Antioch's is. Antioch's constant transport models are
working fine, but the mixture-averaged ones are yielding instabilities.
Still investigating these anomalies.