Rename ElasticMembrane/CableBase to Abstract
To implement Rayleigh damping, we're going to need to reuse the
implementation of element_time_derivative and mass_residual. So, we'll
tuck that in a base class. But not everyone will need it, so we'll put
the current Base in Abstract and the stuff that doesn't need it we can
tuck in Abstract.