Add Rayleigh damping
This adds Rayleigh damping for ElasticCable and ElasticMembrane Physics
classes. The mass_residual we were able to reuse, but the element_time_derivative
not so much. There's still a fair bit of code reuse we ought to be able
to get out of this, but pressing on for now. Note that for the Jacobians, we
are missing derivatives of the elasticity tenstor; of those currently
available, these terms will only be nonzero for the hyperelasticity models.
The contributions look small in testing, but not small enough to pass
numerical vs. analytical Jacobian testing.