Convert more tests away from configure dependence
These tests required reading a mesh and one test additionally required
a restart file. This required adding AC_CONFIG_LINKS calls to setup
symlinks. This is still not an ideal solution, but it scales much
better than requiring it for every test. Currently, there's only a
handful of these. We'll need a few more for other input files.