Update thermo-chem dependent tests to be configure independent
This sets up symlinks to the necessary input files for the
thermochemistry libraries in the tests and then removes all the
configure business for the .sh files. Now the .sh files know where to
look as well as creating symlinks to the input_files that are needed so
each of the input files can locally reference input_files in the build
directory. Thus, we don't need configure business for those input files
anymore either.