Fix make dist
This updates the build system to get `make dist` working with the
test restructuring. The main PITA here is that AC_CONFIG_LINKS sets
up the linked file for distribution (good). BUT, if that file is part
of a directory that you add to EXTRA_DIST, the AC_CONFIG_LINKS'ed file
gets *removed* from the distribution (bad). I don't understand why.
To work around it, then, we build up the list of files we'll want to
distribute in the configure.ac variables. Then, we use those variables
in the top-level Makefile.am.

As of this commit, `make distcheck` passes.