Check T^2 in AntiochEvaluator::check_and_reset_temp_cache
Before, we were checking just T. However, in both the cache and
the argument coming in, T is a reference. So, if T changed upstream,
then the expression evals to true even though we need to rebuild
the TempCache. So, we'll check the next cheapest thing, which is T^2.
This didn't bite us in the PDE solves because we actually had different
addresses for T when constructing this, but it would produce incorrect
results from the thermo apps. Picked up on this in a separate branch
where I'm adding CppUnit-based tests of the thermochemistry interfaces.