Go back to GenericVariable:<physics_name> section
For GenericVariables, we'd originally had this naming scheme for
the subsections of the [Variable] section, but I thought it was
too redundant. Well, we need a factory for VariableBase now and
I don't see a way of doing that without being able to extract the
fact that we have a generic variable. So we're going back to the
more redundant naming.