Cache a GetPot pointer in MultiphysicsSystem
The new boundary condition construction will be dropped in
MultiphysicsSystem::init_data (so we can avoid all the crap we
were doing before of caching things at the beginning to wait until
we had a system), but we still need a GetPot object. So we'll cache
a GetPot pointer to use during init_data.