Update axisym reacting test
Had erroenously put parabolic profile as a function of y and not x.
It shoudln't really matter, except that the velocity solution will
depend on what order you add the no slip boundary and the inlet profile.
Making this test a proper parabolic profile on the inlet should fix
that. So have updated the input file and the gold data for this test.