Add notion of Neumann BC sign to FEVariablesBase
Depending on the Physics, the Neumann boundary condition term in the
weak form could have a positive or negative sign. For example, when
written in the form M\dot{u} = F(u), the Heat equation generates
a '+' when integrated by parts, while Maxwell's equations generate
a '-'. So, we'll let the Physics declare whether it should be '+'
or '-' for that variable. This can then be used by the boundary
condition evaluation when evaluating the Neumann boundary conditions.