Set Doxygen INLINE_INHERITED_MEMB  = NO
I've come to really prefer this layout, where now methods from that
base class that are not overridden are hidden behind a tab. So we
can still easily find them on the same page, but it significantly
declutters looking for methods in the current object, especially
for things like Physics subclasses.