CatalyticWallBase now subclasses NeumannBCAbstract
CatalyticWallBase and its subclasses now subclass NeumannBCAbstract.
Introduced new constructors for forward compatibility for constructing
these objects using NeumannBCFactory subclasses. The idea is that
the Factory now will deal with grabbing the species variable indices
from the System, as well as all the parsing. We also will use SharedPtr
for the Chemistry and Catalycity objects. This makes these objects
more self-contained and the interface more clear (and we can get
rid of the clone() in the CatalycityBase). Marked all deprecated
methods; these will be removed once we move to the new boundary
condition construction paradigm.