Refactor Variable classes
This now refactors the Variable classes so that there's only one
base class: FEVariablesBase. Thus, all the relevant content that was in
the VariablesBase subclasses (e.g. u_var() methods) have been moved to
the FEVariablesBase subclasses. We also use the new intermediate
subclasses, where appropriate. While we've removed the VariablesBase
classes from the build system, we haven't removed the source yet,
we'll do that in a later commit.

The primary purpose of this is 1. It makes more sense and 2. The split
was introduced in the beginning because of the need for in BCHandling
subclasses to have access to Variable information, but not
having/needing the info for parsing the FE info. But now, we can just
use the VariableWarehouse to fetch the Variables that we need (which is
what we updated in a previous commit).  Then, we have this single
hierarchy for the Variable information, which is much cleaner.