Update regression/thermally_driven_2d_flow.sh test
This removes the 'general_heat_flux' boundary condition and replaces
it with adiabatic. Note that there's no resulting change in the
solution, since supplied heat flux was actually 0.0. This originally
was meant as a unit test for the BC factory stuff, but it is being
completely redone downstream, so let's get rid of it here since there's
an appropriate substitute. Note also that this test now uses the
"generic_solution_regression" instead of test_thermally_driven_flow.
It's still not as good as running through grins first, but it's a
step in the right direction.