Pin v component in penalty_poiseuille tests
Previously, the penalty_poiseuille tests were not setting the
Dirichlet value for the v-component of velocity. This led to
small, but detectable, changes in the solution values downstream
in grinsfem/grins#403 when the component was getting pinned. Since
that's the correct thing to do anyway, we change it here. This
required updating the gold data test/test_data/penalty_poiseuille.xdr
and test/test_data/penalty_poiseuille_stab.xdr.