Add BuilderHelper::parse_var_sections_vector
Also update parse_var_sections to use parse_var_sections_vector.
We need this function to preserve the ordering of the Variables
subsections so that when we parse them in order to add the Variables
to the System, we give the user some control over the ordering
that goes into the System. We preserve the old std:;set version
since the BoundaryCondition construction doesn't care.

Note that the search for duplicates is now O(n), where n is the
number of Variable subsections, instead of O(log(n)) (so over
all the subsections, it's O(n^2) instead of O(nlog(n)).
I don't think this is a big deal since the number of Variable
subsections shouldn't ever really be that large. Nevertheless,
I wanted to note it here in case someday it's proven I was grossly
wrong.