Update Variable order in some test input files
Right now, there is no difference in behavior. In the subsequent
Variable construction changes, this is the order that the
variables will be added to the system. As such, we want to preserve
the ordering from the previous tests for 1. Solver behavior
consistency and 2. Consistency with the variable ordering in the
gold data.