Introduce SingleVariable, move subclasses
This introduces the FEVariablesBase subclass SingleVariable. It's
main purpose to be a parent class to syntatic-sugar subclasses that
sanity checks it's really a single variable. Went ahead and moved
the single variable subclasses to the same header, no sense in having
them separate. The single variables that are intended to be SCALAR
will get their own subclass.