Add regression/ozone_flame_antioch_kinetics_theory.sh test
This is currently an XFAIL test as there is (likely) some bug(s) in
the Antioch KineticsTheory functionality. This test is restarting
from the constant transport version; currently, the residual NANs.
Once we can get this working, we'll also need to update the gold
data --- currently we're using the constant transport as a placeholder.