Call MultiphysicSystem::read_input_options() in testing helper
For unit tests that may be using "raw" MultiphysicsSystem (i.e. outside
of the Simulation class), this is one of the necessary steps. In
particular, this function caches a pointer to the GetPot object
that will be subsequently passed to things like the boundary condition
factories. So, let's make sure and call it when the developer is
using this testing helper function.