Add QoIOutput
Class to manage the different stages of QoI output. Currently,
just have display and file output. Actual implementation of the
output is delegated to the specific QoI (for which the base class
has a default implementation); this class just manages what to
call when and some pretty printing around the display output.