Add Runner class
This encapsulates initialization, construction, and running of
a Simulation object and accompanying data structures. The idea is that
this should make 1. GRINS main() a two-liner program and 2. Make it
much easier for users to make their own GRINS-based programs, e.g.
they instantiate their factory object and then setup and run the Runner;
that's it.