Move unused var checking to Runner
Now that we have a Runner class, we now move the check for GetPot
UFOs from the Simulation constructor to Runner::run(). This way,
if the user adds command line variables of the form var='option',
they can grab the inputfile and use the input.have_variable("option")
trick to not trigger the UFO error. This will also let them do other
things like add their own options to the input file and parse them
after Runner sets up and not trigger the UFO error.