Distribute src files needed for instantiation
These files are explicitly listed in the libgrins_SOURCES to avoid
compiling them twice - once for compiling the .C and once when they're
included for instantiation. Instead, we just include them for
instantiation so they're only compiled onced. But, that means we need
to make sure to add them to the data for install/distribution