Fix ozone_flame efficiencies specification
Antioch recognized this, but this was not compatible with Cantera
so Cantera was doing the incorrect kinetics. This fixes that, but
there are still other problems (negative temperatures going to Cantera)
that prevent the example running with Cantera out-of-the-box.

Thanks to @klbudzin for pointing this out to me in meatspace.