Start parsing species names from XML file when using Cantera
We'll eventually move Antioch parsing in this direction, but that's
more work since there's different kinds of parsers, so let's do the
easy one first.

For now, we hack around the VariableTest by just setting the
thermochemistry_library as 'antioch', which at the moment will
just fallback to parsing the species from the GRINS input file.
We'll need to update these particular unit tests in more depth
once Antioch parsing is updated in a similar fashion.