Add ASCII, XML support to build_reaction_set
In reality, both formats are using XML to parse the kinetics, but
the ASCII format requires an additional flag for the filename for
the kinetics data whereas XML it will all be stashed in one place.

This commit is also leveraging a new Antioch API that was part of
libantioch/antioch#261. Thus, the minimum Antioch hash that will
be required after this commit is  libantioch/antioch@7cebf19.

We'll work on ChemKinParser support later.