Add XML build_reaction_set unit test for air9sp
Just like the build_species_name test, we also check a different
mixture from the same XML file. It's the same kinetics section, but
it *should* include more of the reactions (but not all) since
we have more species in the mixture, so we test all the reactions
individually, at least the type and number and efficiencies (if
three bodY), but not each set of coefficients.