Update ozone flame test
Since the three body efficiencies were totally wrong, we need to update
the files and the gold standard. This yield identical output to the just
updated example. I used the parallel testing script and check this test
on 1--16 processors. To get convergence on 15 and 16 processors, I had
to strengthen the linear solver preconditioner. Finally, I don't know if
there was some weird cancellation going on or what, but if I turn *on*
the clip negative density option, this test fails to converge. O_O

Fun fact, I could get the old restart
to converge using the new files, but took like 5 mins and 70 something
Newton steps... hoorary line search!