Add AMR step to run.sh for Antioch case
This will do 5 adaptive refinement steps using a KellyErrorEstimator
to estimate the error. Note that we don't bother converging the
nonlinear solver completely in between since we just want it to be
a good enough initial guess for the next AMR step. Note also we
don't converge the final step probably because of poor scaling of
the equations.