Get Cantera regression test actually working
LONG overdue. This is the 2D nitridation recombination dummy that
we have for like 1093841298 other Antioch tests. Had to soften the
N fraction vs. Antioch because got negative temp during Newton solve
that Cantera complained about (we're not trimming low temps like
we are for Antioch. That will change in the interface in the GRINS
interface in the future).