Error out Rayleigh Damping if using 1st order time solver
We haven't updated the RayleighDamping Physics classes to support
the second_order_var interface so this will produce the incorrect
thing if we're using a libMesh::FirstOrderUnsteadySolver so we'll
error out for now until we can update to support this interface. Not
entirely trivial just because we'll need to deal with two different
Jacobians.