Make setting non-specified dirichlet BC components zero optional
Right now, the user only needs to specify the value of a subset of
variable components and the BC factory will make all the other
components zero. There are some cases however where we may not want to
do this. So this first step is making it optional. We do this through
a constructor argument in the base class and then the subclass can
dictate whether or not to set this new argument true or false.