template CartesianSolidMechanics,HyperelasticityBase on Dim
Not done yet, but what we're doing is templating on the dimension
so we can deal with plane strain 2D and full 3D with the same code
and have the compiler pick the correct code paths at compile time.
We haven't actually updated the core code to do this yet, but that
will be the next step.