Use default destructor
The neat thing about std::unique_ptr is that it will automatically
delete the pointer for you in its destructor.
And the neat thing about std::vector
is that its destructor will loop over all its elements and call
each of their destructors. So with a std::vector<std::unique_ptr>,
you don't need to do anything in the destructor, it all just works!