Make QoIBase::init_context pure virtual, add impls
libMesh requires us to specify what we want from the FEMContext,
so that means that the interface we use require subclasses to
actually use init_context to say what they need. If we'd done this
originally, this could have save a lot of time.

Added implementations that were missing, along with documentation,
including some todos on how we might redo someof the laser absorption
and other cousins/subclasses of IntegratedFunction.