Update cppunit.m4
Newer versions of CppUnit do not include the cppunit-config.
Libmesh is using macros that require one of these newer versions
anyway so I don't even bother checking for it anymore. Also
took out the version checks for simplicity. If we need to bifurcate
on version, we can do something about it later. But really, I'd
rather just move to catch2 one of these days.